+++ tests/testthat/test_violin_plot.R
@@ -3,21 +3,21 @@ G<-igraph::upgrade_graph(karate)
 
 #Network summary plot
 test_that("violin plots with pre-specified subsample size",
-          {expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G,sparse = FALSE), subsample_sizes = 20,
-                  max_cycle_order = 7, save.plot = FALSE)})}
+          {suppressMessages(expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G,sparse = FALSE), subsample_sizes = 20,
+                  max_cycle_order = 7, save.plot = FALSE)}))}
 )
 
 test_that("violin plots with auto-selected subsample size",
-          {expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G,sparse = FALSE), Ns=5,
-                             max_cycle_order = 7, save.plot = FALSE)})}
+          {suppressMessages(expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G,sparse = FALSE), Ns=5,
+                             max_cycle_order = 7, save.plot = FALSE)}))}
 )
 
 test_that("violin plots with igraph object",
-          {expect_no_error({violin_netsummary(G, Ns=5,
-                             max_cycle_order = 7, save.plot = FALSE)})}
+          {suppressMessages(expect_no_error({violin_netsummary(G, Ns=5,
+                             max_cycle_order = 7, save.plot = FALSE)}))}
 )
 
 test_that("violin plots with sparse matrix object",
-          {expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G), Ns=5,
-                             max_cycle_order = 7, save.plot = FALSE)})}
+          {suppressMessages(expect_no_error({violin_netsummary(igraph::as_adjacency_matrix(G), Ns=5,
+                             max_cycle_order = 7, save.plot = FALSE)}))}
 )